Fixed tooltip visibility
